Move changelog functions to separate file
