+++ src/commcare_cloud/manage_commcare_cloud/manage_commcare_cloud.py
@@ -7,7 +7,8 @@ from commcare_cloud.manage_commcare_cloud.configure import Configure
 from commcare_cloud.manage_commcare_cloud.get import Get
 from commcare_cloud.manage_commcare_cloud.list_vault_keys import ListVaultKeys
 from commcare_cloud.manage_commcare_cloud.install import Install
-from commcare_cloud.manage_commcare_cloud.make_docs import MakeDocs, MakeChangelog
+from commcare_cloud.manage_commcare_cloud.make_docs import MakeDocs
+from commcare_cloud.manage_commcare_cloud.make_changelog import MakeChangelog
 
 COMMAND_TYPES = [
     Install,
